Revert "New modal"
