+++ challenges.html
@@ -12,14 +12,12 @@
       property="og:image"
       content="https://nelsonthebrave.github.io/hacker-escape-rooms/assets/image 2large.png"
     />
-    <link rel="stylesheet" href="assets/main.css"  />
+    <link rel="stylesheet" href="assets/main.css" />
     <script
       src="https://kit.fontawesome.com/c866a2c3b1.js"
       crossorigin="anonymous"
     ></script>
-    <script defer src="main.js" type="module"></script>
-    <script defer src="challenges.js" type="module"></script>
-    
+    <script defer src="challenges.js"></script>
     <title>[ESC] Hacker Escape Rooms</title>
   </head>
 
@@ -159,6 +157,6 @@
         <li><a href="">Twitter</a></li>
       </ul>
     </footer>
-    
+    <script src="main.js"></script>
   </body>
 </html>
