+++ index.html
@@ -6,13 +6,13 @@
     <meta property="og:description" content="Play challenging and entertaining Escape games with friends, either online or at our venue"/>
     <meta property="og:type" content="Website"/>
     <meta property="og:image" content="https://nelsonthebrave.github.io/hacker-escape-rooms/assets/image 2large.png"/>
-    <link rel="stylesheet" href="assets/main.css"  />
+    <link rel="stylesheet" href="assets/main.css" />
     <script
     src="https://kit.fontawesome.com/c866a2c3b1.js"
     crossorigin="anonymous"
   ></script>
-    <script defer src="challenges.js" type="module"></script>
-    <script defer src="main.js" type="module"></script>
+    <script defer src="main.js"></script>
+    <script defer src="challenges.js"></script>
     <title>[ESC] Hacker Escape Rooms</title>
   </head>
 
@@ -184,7 +184,5 @@
         </ul>
     </footer>
 
-    
-    
   </body>
 </html>
