+++ src/main.sass
@@ -73,9 +73,4 @@ html
 @import 'sections/info' 
 @import 'sections/footer'
 @import 'sections/filter'
-
-
 @import 'sections/challenges-site'
-
-// modal
-@import 'sections/booking-modal'
