--- src/sections/booking-modal.scss
@@ -1,9 +1,0 @@
-
-
-.bookingSceneContainer__SectionClass { 
-    background-color: red;
-    color: red;
-    padding: 0px;
-    }
-
-
